gui-cql/requirements.txt
fastapi
uvicorn[standard]
cassandra-driver
jinja2
python-multipart


